# (perf): disable org-agenda while building the cache