@natask 2020-09-28
the core issue is that org-roam opens files using `find-file-noselect` when updating db, enabling whatever else the user has configured to run when files are open sans what org-roam ends up disabling.
 we should move towards a whitelist instead of a blacklist of executed commands by replacing `find-file-noselect` with `with-temp-buffer` + `insert-file-contents`. Then enabling functions to run as needed will alleviate future issues.