@juergenhoetzel 2020-09-29
> `org-roam-db-build-cache` has always read the file contents into a temp buffer, not sure why it's affected by `org-agenda`, so I'd like to do a deeper investigation before merging this.

`org-mode` itself triggers the building of `agenda-files-menu`:

```
(define-derived-mode org-mode outline-mode "Org"
...
  (org-load-modules-maybe)
  (org-install-agenda-files-menu)
...

```