@natask 2020-09-29
you are right. my bad for not checking closely.
this is why it is affected (org-agenda). 
https://github.com/org-roam/org-roam/blob/be64f107e9d24cb457a4fb5d92e8aed26b3d06e7/org-roam-macs.el#L60-L70 in function https://github.com/org-roam/org-roam/blob/be64f107e9d24cb457a4fb5d92e8aed26b3d06e7/org-roam-macs.el#L49-L64