@jethrokuan 2020-09-29
The correct solution is to design a version of `(org-mode)` that only enables the things that are needed (setting regexps, outline-level etc.).

I'll accept this as a hotfix, though.