@natask 2020-09-29
you might want to propagate the hot-fix to `org-roam--with-temp-buffer` so that other usages of the function outside `org-roam-db-build-cache` can benefit from it as well.